Add files via upload
IoT based Smart shoe system for the blind is a system made with the help of ultrasonic
sensors paired to an Arduino UNO board. Internet of things is all about making physical objects
communicate with other objects or even with humans. It is an enabling technology which has a rapid
development and growth in the market. In our India there are almost 40 million blind people among
which 1.6 million are children. Blind people face great difficulty to travel independently. They have
to depend on others in many aspects of their life. The Major problem is when they walk on the road.
With a stick in hand, they cannot detect every obstacle that comes in their way. The Smart shoe
design provides a long- term solution for the blind to walk on roads independently. The smart
shoe will help the Blind person to reach his destination independently. It is built using IoT Technology
in which the shoe will be embedded with various sensors, Microcontroller and buzzers. The shoe
warns the user by making noise with the buzzer when he/she walks in front of an obstacle. To improve
the efficiency, Smart glasses are designed using IoT which is also embedded with sensors and helps in
detecting the objects by covering a larger area. The smart shoe and the smart glasses communicate
and coordinate with each other to ensure that the user does not collide with any obstacle in his way